✨ feat(entity): 添加 SysDept 实体及其仓库支持
✨ feat(repository): 增加 SysRoleRepository 以支持角色相关查询

🐛 fix(service): 修复 SysUserServiceImpl 中的用户查询逻辑

🔧 chore(entity): 更改 SysUserRole 和 SysRole 使用 @Getter 和 @Setter 注解

🔧 chore(entity): 更新 SysDept 使用 @Getter 和 @Setter 注解

+++ src/main/java/top/werls/vben/system/repository/SysRoleRepository.java
@@ -27,6 +27,6 @@ public interface SysRoleRepository extends CrudRepository<SysRole, Integer> {
     List<SysRole> findByUserId(Long userId);
 
     @Query("select u from SysUser u  left  join SysUserRole ur on u.uid =ur.uid where ur.rid = ?1")
-    List<SysUser> findByRoleId(Integer rid);
+    List<SysUser> findByRoleId(Long rid);
 
 }
